Reorganise where the changed tab property lives in the context.
