+++ uSync.Migrations/Context/SyncMigrationContext.cs
@@ -48,9 +48,6 @@ public class SyncMigrationContext : IDisposable
     private HashSet<string> _blockedTypes = new(StringComparer.OrdinalIgnoreCase);
     private Dictionary<int, Guid> _idKeyMap { get; set; } = new();
 
-    // tabs that are to be changed
-    private List<TabOptions> _changedTabs { get; set; } = new List<TabOptions>();
-
     /// <summary>
     ///  is this item blocked based on alias and type. 
     /// </summary>
@@ -74,15 +71,7 @@ public class SyncMigrationContext : IDisposable
     /// </summary>
     public Guid GetKey(int id)
         => _idKeyMap?.TryGetValue(id, out var key) == true ? key : Guid.Empty;
-    /// <summary>
-    /// Add changed tabs to the context.
-    /// </summary>
-    public void AddChangedTabs(TabOptions tab)
-        => _changedTabs.Add(tab);
-
-    public List<TabOptions> GetChangedTabs()
-        => _changedTabs;
-    
+   
     public void Dispose()
     { }
